+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Change Log
 ==========
 
+- 新增 BadRequest 與 UserError 類別。
+
 ### 0.0.5 (2017-07-03)
 
 - 新增 MonologFluentHandler 類別